move tests into their own package
Tests should not figure inside the main client's package, they should instead be developed independantly
