+++ tests/utils/string.py
